requirements_dev.txt
-r requirements_base.txt
coverage
django-debug-toolbar
django-extensions
factory-boy
Faker
selenium
